pread uses size_t while	read uses unsigned int.

+++ src/logging.cc
@@ -191,13 +191,13 @@ GLOG_DEFINE_string(log_backtrace_at, "",
 #include <BaseTsd.h>
 #define ssize_t SSIZE_T
 #endif
-static ssize_t pread(int fd, void* buf, unsigned int count, off_t offset) {
+static ssize_t pread(int fd, void* buf, size_t count, off_t offset) {
   off_t orig_offset = lseek(fd, 0, SEEK_CUR);
   if (orig_offset == (off_t)-1)
     return -1;
   if (lseek(fd, offset, SEEK_CUR) == (off_t)-1)
     return -1;
-  ssize_t len = read(fd, buf, count);
+  ssize_t len = read(fd, buf, (unsigned int)count);
   if (len < 0)
     return len;
   if (lseek(fd, orig_offset, SEEK_SET) == (off_t)-1)
